yge/requirements.txt
Flask
Redis
Tensorflow
wget
Pillow
numpy
cassandra-driver